+++ test/main.js
@@ -6,7 +6,7 @@ const { modifySettings } = require('./modifySettingsHelper')
 
 chai.should()
 chai.use(chaiAsPromised)
-const timeout = process.env.CI ? 30000 : 10000
+const timeout = process.env.CI ? 60000 : 10000
 
 describe('stretchly', function () {
   this.timeout(timeout)
@@ -76,6 +76,27 @@ describe('stretchly', function () {
               .waitUntilWindowLoaded()
               .getWindowCount().should.eventually.equal(1)
               .windowByIndex(0).browserWindow.isVisible().should.eventually.be.false
+          })
+      })
+  })
+
+  it('main window is stretchly ', function () {
+    return modifySettings('isFirstRun', false)
+      .then(() => {
+        this.app = new Application({
+          path: electronPath,
+          args: [
+            `${__dirname}/../app`
+          ]
+        })
+
+        return this.app.start()
+          .then(() => {
+            chaiAsPromised.transferPromiseness = this.app.transferPromiseness
+            return this.app.client
+              .waitUntilWindowLoaded()
+              .windowByIndex(0).browserWindow
+              .getTitle().should.eventually.equal('stretchly')
           })
       })
   })